Updated tests.
Updated unit tests for batching with a GPJax Dataset.
