+++ tests/test_abstractions.py
@@ -1,6 +1,3 @@
-import typing as tp
-
-import jax
 import jax.numpy as jnp
 import jax.random as jr
 import optax
@@ -10,24 +7,9 @@ from jax.experimental import optimizers
 
 import gpjax as gpx
 from gpjax import RBF, Dataset, Gaussian, Prior, initialise, transform
-from gpjax.abstractions import batch_loader, fit, fit_batches, optax_fit
+from gpjax.abstractions import fit, fit_batches, optax_fit
 
 tfd = tf.data
-
-
-def _dataset_to_tf(dataset, prefetch_buffer=1, batch_size=32):
-    X, y, n = dataset.X, dataset.y, dataset.n
-
-    batch_size = min(batch_size, n)
-
-    # Make dataloader, set batch size and prefetch buffer:
-    ds = tfd.Dataset.from_tensor_slices((X, y))
-    ds = ds.cache()
-    ds = ds.repeat()
-    ds = ds.shuffle(n)
-    ds = ds.batch(batch_size)
-    ds = ds.prefetch(prefetch_buffer)
-    return ds
 
 
 @pytest.mark.parametrize("n", [20])
@@ -86,40 +68,6 @@ def test_stop_grads(optim_pkg):
     assert learned_params["x"] != params["x"]
 
 
-@pytest.mark.parametrize("batch_size", [1, 10])
-@pytest.mark.parametrize("n", [50, 100])
-def test_batcher(batch_size, n):
-    x = jnp.linspace(-3.0, 3.0, num=n).reshape(-1, 1)
-    y = jnp.sin(x)
-    D = Dataset(X=x, y=y)
-    D = _dataset_to_tf(D, batch_size=batch_size)
-    batcher = batch_loader(data=D)
-    Db = batcher()
-    assert Db.X.shape[0] == batch_size
-    assert Db.y.shape[0] == batch_size
-    assert Db.n == batch_size
-    assert isinstance(Db, Dataset)
-
-    Db2 = batcher()
-    assert any(Db2.X != Db.X)
-    assert any(Db2.y != Db.y)
-    assert Db2.n == batch_size
-    assert isinstance(Db2, Dataset)
-
-
-@pytest.mark.parametrize("nb", [20, 50])
-@pytest.mark.parametrize("ndata", [10])
-def test_min_batch(nb, ndata):
-    x = jnp.linspace(-3.0, 3.0, num=ndata).reshape(-1, 1)
-    y = jnp.sin(x)
-    D = Dataset(X=x, y=y)
-    D = _dataset_to_tf(D, batch_size=nb)
-    batcher = batch_loader(data=D)
-
-    Db = batcher()
-    assert Db.X.shape[0] == ndata
-    assert isinstance(batcher, tp.Callable)
-
 
 @pytest.mark.parametrize("nb", [20, 50])
 @pytest.mark.parametrize("ndata", [50])
@@ -137,18 +85,23 @@ def test_batch_fitting(nb, ndata, jit_compile):
     params, trainable_status, constrainer, unconstrainer = initialise(svgp)
     params = gpx.transform(params, unconstrainer)
     objective = svgp.elbo(D, constrainer)
-    D = _dataset_to_tf(dataset=D, batch_size=nb)
-    batcher = batch_loader(data=D)
-    pre_mll_val = objective(params, batcher())
+
+    D = Dataset(X=x, y=y)
+    D = D.cache()
+    D = D.repeat()
+    D = D.shuffle(D.n)
+    D = D.batch(batch_size=32)
+    D = D.prefetch(buffer_size=1)
+
     opt_init, opt_update, get_params = optimizers.adam(step_size=0.1)
     optimised_params = fit_batches(
         objective,
         params,
         trainable_status,
+        D,
         opt_init,
         opt_update,
         get_params,
-        get_batch=batcher,
         n_iters=5,
         jit_compile=jit_compile,
     )
