+++ tests/test_types.py
@@ -1,5 +1,6 @@
 import jax.numpy as jnp
 import pytest
+import typing as tp
 
 from gpjax.types import Dataset, NoneType, verify_dataset
 
@@ -35,3 +36,43 @@ def test_y_none():
     d = Dataset(X=x)
     verify_dataset(d)
     assert d.y is None
+
+
+
+@pytest.mark.parametrize("batch_size", [1, 10])
+@pytest.mark.parametrize("n", [50, 100])
+def test_batcher(batch_size, n):
+    x = jnp.linspace(-3.0, 3.0, num=n).reshape(-1, 1)
+    y = jnp.sin(x)
+    D = Dataset(X=x, y=y)
+    D = D.cache()
+    D = D.repeat()
+    D = D.shuffle(D.n)
+    D = D.batch(batch_size)
+    D = D.prefetch(buffer_size=1)
+    batcher = D.get_batches()
+    Db = batcher()
+    assert Db.X.shape[0] == batch_size
+    assert Db.y.shape[0] == batch_size
+    assert Db.n == batch_size
+    assert isinstance(Db, Dataset)
+
+    Db2 = batcher()
+    assert any(Db2.X != Db.X)
+    assert any(Db2.y != Db.y)
+    assert Db2.n == batch_size
+    assert isinstance(Db2, Dataset)
+
+
+@pytest.mark.parametrize("nb", [20, 50])
+@pytest.mark.parametrize("ndata", [10])
+def test_min_batch(nb, ndata):
+    x = jnp.linspace(-3.0, 3.0, num=ndata).reshape(-1, 1)
+    y = jnp.sin(x)
+    D = Dataset(X=x, y=y)
+    D = D.batch(batch_size=nb)
+    batcher = D.get_batches()
+
+    Db = batcher()
+    assert Db.X.shape[0] == ndata
+    assert isinstance(batcher, tp.Callable)
